rename enter coordinates window
