+++ oceannavigator/frontend/src/components/Map/drawing.js
