+++ oceannavigator/frontend/src/components/Map/utils.js
