+++ oceannavigator/frontend/src/components/ModifyFeaturesWindow/FeatureCard.jsx
