+++ oceannavigator/frontend/src/stylesheets/components/_ModifyFeaturesWindow.scss
@@ -1,4 +1,4 @@
-.EnterCoordsWindow {
+.ModifyFeaturesWindow {
   .feature-col {
     max-height: 75vh;
     overflow-y: scroll;
