+++ oceannavigator/frontend/src/stylesheets/components/_all.scss
@@ -10,7 +10,7 @@
 @import "DatasetDropdown";
 @import "DatasetSelector";
 @import "DrawingTools";
-@import "EnterCoordinatesWindow";
+@import "ModifyFeaturesWindow";
 @import "FeatureCard";
 @import "ImageSize";
 @import "InfoHelpWindow";
